Add AI fallbacks, async bootstrap & UI cleanups
Introduce on-device AI page context tracking and local fallbacks: catalog, pricing, suppliers, companions and forecasting handlers, plus NotificationCenter hooks and a command to apply catalog filters. Replace several await availability checks with synchronous calls. Major AppCore refactor: initialize DB and services off the main thread via Task.detached, make service properties optional until ready, return bootstrap results to the main actor, run companion auto-discovery in background, and add AppCoreError with databasePath now throwing. Make login and seedFirstAdmin async and perform auth calls off the main thread. Miscellaneous fixes and cleanups: remove numerous platform #if blocks for consistent UI, improve geofence/location handling and error reporting in views, add new warehouse/QR model files, and other small service/model updates.

+++ Weird Parts IOS/Weird Parts IOS/Features/Warehouse/IOSInventoryGridPage.swift
@@ -19,7 +19,7 @@ struct IOSInventoryGridPage: View {
     @State private var selectedFilter: StockFilter?
     @State private var actionError: String?
     @AppStorage("lastInventoryLocationType") private var lastLocationType = "warehouse"
-    @AppStorage("lastInventoryLocationId") private var lastLocationId: Int64 = 1
+    @AppStorage("lastInventoryLocationId") private var lastLocationId: Int = 1
 
     private struct LocationOption: Identifiable, Hashable {
         let locationType: String
@@ -80,12 +80,12 @@ struct IOSInventoryGridPage: View {
         }
         .onChange(of: selectedLocationId) {
             lastLocationType = selectedLocationType
-            lastLocationId = selectedLocationId
+            lastLocationId = Int(selectedLocationId)
             loadData()
         }
         .task {
             selectedLocationType = lastLocationType
-            selectedLocationId = lastLocationId
+            selectedLocationId = Int64(lastLocationId)
             loadLocations()
             loadData()
         }
